+++ application/models/Site_Settings_Model.php
@@ -17,37 +17,48 @@ class Site_Settings_Model extends CI_Model
 	}
 
 	function update_site_settings($images) {
-		// var_dump($this->input->post());
-		// echo "<br>";
-		// in_array ile gelen alanlar varmı onu ogrenip alanlar var ise bu alanların degerini dbye yaz yok ise alanın degerini boşalt.
+
 		$db_rows = $this->get_site_settings();
-		foreach ($this->input->post() as $post_key => $post_value) {
-			
-		}
 		foreach ($db_rows as $db_key => $db_value) {
-				if ($db_value->settings_name == $post_key) {
-					echo "db_name: ".$db_value->settings_name.", post_key: ".$post_key.", post_value: ".$post_value."<br>";
-				}
-			}
-		exit;
+			$site_settings_update_data = array(
+				'settings_value' => ''
+			);
+			$this->db->set($site_settings_update_data);
+			$this->db->where('settings_name', $db_value->settings_name);
+			$this->db->update('site_settings');
+		}
+		foreach ($this->input->post() as $key => $value) {
+			$site_settings_update_data = array(
+				'settings_value' => $value
+			);
+			$this->db->set($site_settings_update_data);
+			$this->db->where('settings_name', $key);
+			$this->db->update('site_settings');	
+		}
 		$this->load->helper('rand_helper');
 		foreach ($images as $key => $value) {
 			if (!empty($value['name'])) {
-				unlink(FCPATH.$this->input->post('old_'.$key));
+				if (!empty($this->input->post('old_'.$key))) {
+					unlink(FCPATH.$this->input->post('old_'.$key));
+				}
 				$randomString = generateRandomString(14);
 				$db_img_name[$key] = 'assets/uploads/system/images/'.$key.'_'.$randomString.'-'.$this->changeName($value['name']);
 			} else {
 				$db_img_name[$key] = $this->input->post('old_'.$key);
 			}
+			if ($this->input->post('delete_'.$key)) {
+				unlink(FCPATH.$this->input->post('delete_'.$key));
+				$db_img_name[$key] = '';
+			}
 			move_uploaded_file($value["tmp_name"], FCPATH.$db_img_name[$key]);
+			$site_settings_update_data = array(
+				'settings_value' => $db_img_name[$key]
+			);
+			$this->db->set($site_settings_update_data);
+			$this->db->where('settings_name', $key);
+			$this->db->update('site_settings');	
 		}
-		$site_settings_update_data = array(
-			'settings_key' => '',
-			'settings_value' => ''
-		);
-		$this->db->set($site_settings_update_data);
-		$this->db->where('settings_name', $this->input->post('id'));
-		$this->db->update('site_settings');
+		
 	}
 
 	public function changeName($change_name){
